+++ eslint.config.mjs
@@ -1,0 +1,16 @@
+import eslint from '@eslint/js'
+import { defineConfig } from 'eslint/config'
+import tseslint from 'typescript-eslint'
+
+export default defineConfig(
+  {
+    ignores: ['dist', 'test-build', '.nyc_output', 'node_modules', 'resources', 'test', 'test_resources'],
+  },
+  eslint.configs.recommended,
+  tseslint.configs.recommended,
+  {
+    rules: {
+      '@typescript-eslint/no-explicit-any': 'error',
+    },
+  },
+)